requirements.txt
attrs==19.1.0
backports.csv==1.0.7
certifi==2019.3.9
chardet==3.0.4
decorator==4.4.0
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.0
django-ckeditor==5.7.0
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.2.0
django-js-asset==1.2.2
django-ranged-response==0.2.0
django-reversion==3.0.3
django-simple-captcha==0.5.11
et-xmlfile==1.0.1
future==0.15.2
httplib2==0.9.2
idna==2.8
ipython-genutils==0.2.0
jdcal==1.4.1
jsonschema==3.0.1
jupyter-core==4.4.0
mysqlclient==1.4.2.post1
nbformat==4.4.0
numpy==1.16.3
odfpy==1.4.0
openpyxl==2.6.2
pandas==0.24.2
patsy==0.5.1
Pillow==6.0.0
plotly==3.9.0
plotly-express==0.1.7
PyMySQL==0.9.3
pyrsistent==0.15.1
python-dateutil==2.8.0
pytz==2019.1
pyyaml==5.1
qiniu==7.2.2
requests==2.21.0
retrying==1.3.3
scipy==1.2.1
six==1.10.0
statsmodels==0.9.0
tablib==0.13.0
traitlets==4.3.2
urllib3==1.24.3
xadmin==2.0.1
xlrd==1.2.0
xlwt==1.3.0
